requirements.txt
-e .
aws-cdk.aws-cloudformation~=1.181.1
aws-cdk.aws-cloudwatch~=1.181.1
aws-cdk.aws-codebuild~=1.181.1
aws-cdk.aws-codecommit~=1.181.1
aws-cdk.aws-codedeploy~=1.181.1
aws-cdk.aws-codeguruprofiler~=1.181.1
aws-cdk.aws-codepipeline~=1.181.1
aws-cdk.aws-codepipeline-actions~=1.181.1
aws-cdk.aws-dynamodb~=1.181.1
aws-cdk.aws-ec2~=1.181.1
aws-cdk.aws-iam~=1.181.1
aws-cdk.aws-kms~=1.181.1
aws-cdk.aws-logs~=1.181.1
aws-cdk.aws-s3~=1.181.1
aws-cdk.aws-s3-assets~=1.181.1
aws-cdk.aws-s3-deployment~=1.181.1
aws-cdk.aws-s3-notifications~=1.181.1
aws-cdk.aws-secretsmanager~=1.181.1
aws-cdk.aws-ssm~=1.181.1
aws-cdk.core~=1.181.1
aws-cdk.pipelines~=1.181.1
aws-cdk.aws-redshiftserverless~=1.181.1
boto3~=1.17.97
